pinctrl: meson: meson8: move GPIO_BSD_EN to the CBUS GPIO controller
Configuring the GPIO_BSD_EN pin is done using different registers:
- output high/low is in CBUS PREG_PAD_GPIO0_O[31]
- pull-up/down is in CBUS PAD_PULL_UP_REG2[0]
- output enable is in CBUS PREG_PAD_GPIO0_O[30]
- unknown purpose: CBUS PREG_PAD_GPIO0_O[29] has to be cleared
- unknown purpose: SECBUS2 AO_SECURE_REG0[0] has to be set

This means five out of six bits needed to configure this pin are in the
CBUS region. Additionally one bit has to be configured which is in the
SECBUS2 region (instead of AOBUS region as the other AOBUS pins).

Move the GPIO_BSD_EN pin to the CBUS group and configure the output
high/low, pull-up/down and output enable bits. The two unknown ones will
be hooked up in the future.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>