clk: meson: meson8b: initialize the HDMI PLL registers
Add the reg_sequence to initialize the HDMI PLL with the settings for
a video mode that doesn't require PLL internal clock doubling. These
settings are taken from the 3.10 vendor kernel's driver for the 1296MHz
PLL setting used for the CVBS video modes. This puts the PLL into a
defined state and the Linux kernel can take over.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>