clk: meson: meson8b: Add the mpeg_rtc_osc_sel clock
The first input of the CLK81 clock tree uses the SoC's external
oscillators. By default it's the 24MHz XTAL from which most frequencies
in this SoC are derived. For power-saving purposes there's a mux to
switch the input between the 24MHz XTAL and the 32kHz RTC oscillator.
Add support for that mux add it to the CLK81 clock tree for a better
representation of how the hardware is actually designed.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>