PCI: dwc/intel-gw: use an MMIO regmap for accessing the "app" registers
Use an MMIO regmap for accessing the "app" registers. This makes it
easier to add support for older Lantiq SoCs where these registers need
to be accessed as big-endian instead of little-endian.

Also add a read table because when reading some (but not all)
undocumented registers an external bus abort is generated on xRX200
SoCs. This affects at least PCIE_APP_PMC which only exists on LGM and
newer, while PCIE_APP_AHB_CTRL has been dropped on LGM.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>