PCI: dwc/intel-gw: Add dw_pcie_ops for the Lantiq PCIe controller IP
Add custom .link_up, .read_dbi and .write_dbi hooks for the PCIe
controller IP on Lantiq SoCs.

.link_up has to poll for the PCIE_APP_PHY_SR_PHY_LINK_UP bit and also
check if the PCIE_APP_RC_DR_DLL_UP bit is set in addition to the
"default" link-status checking with the PCIE_PORT_DEBUG1 register.

.read_dbi and .write_dbi need custom implementations because we need to
read/write the values aligned to four-byte addresses from the hardware
in big-endian, which is also the CPU endianness.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>