PCI: dwc/intel-gw: Add the pci_ops for the Lantiq PCIe IP variant
The Lantiq IP variant requires custom pci_ops for both, reading/writing
the controller's own config as well as for reading/writing the child
devices config. The DWC IP revision this controller is based on does not
have any iATU windows defined.

The own config space can only be accessed at addresses which are 4-byte
aligned. Additionally the values are coming back in little-endian
byte-order, which we need to convert to big-endian byte-order to match
the CPU's endianness. The second requirement makes it impossible to use
pci_generic_config_{read,write}32 because it reads the values as
little-endian and then operates on them on a big-endian system as if
they were big-endian. Use dw_pcie_{read,write}_dbi as our
.{read,write}_dbi implementations can handle this case already.

Accessing the child devices config space requires using the "config" IO
register space. Read and write access has to be 4-byte aligned again,
but this time the values are coming back as big-endian already. Use
pci_generic_config_{read,write}32 for accessing the child config space
along with a custom .map_bus hook.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>