MIPS: lantiq: xway: update the clock alias' for the mainline PCIe driver
Update the clock alias' for the PCIe clocks to conform with the
dt-bindings (even though we don't use device-tree yet as the clock code
is smart enough to fall back to the clock lookup table instead). This
makes the PCIe controller driver find the clocks and start working.

The documented clock names are:
- "pcie" for the core clock, this can be empty as well (which is what we
  prefer for backwards compatibility with the pcie-intel-gw driver)
- "pcie_bus" for the PCIe bus clock
- (xRX200 only) "ahb" for the AHB clock

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>