dt-bindings: interrupt-controller: Document the Lantiq MSI PIC
The Lantiq MSI Programmable Interrupt Controller is used on Lantiq
xRX200 and xRX300 SoCs. It takes four to sixteen interrupts from the
parent interrupt-controller (ICU), one for each MSI. It uses two
register areas: "pic" contains the registers for this PIC and "msi"
contains the actual MSI data. A dedicated clock gate has to be enabled
when using this MSI PIC. Add the documentation for this
interrupt-controller.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>