net: dsa: lantiq_gswip: Fix a crash when a fixed-link is configured
Some boards using the Lantiq xRX200 SoC use a fixed-link to connect a
GSWIP port to another device. Known examples are:
- some AVM Fritz devices (including not limited to 7490) have a separate
  QCA SoC for offloading wifi. Communication with that other SoC uses
  port 5 on the GSWIP with a fixed Gbit/s link
- EasyBox 904 xDSL connects GSWIP port 0 to a RTL8367 switch and port 5
  to a Ralink SoC for wifi offloading

Only configure the GSWIP_MDIO_PHYp(port) register when the PHY is used
for negotiating the link. MDIO auto polling also can be kept disabled
if we don't have a PHY connected to that port. Enabling the port in the
GSWIP_MII_CFG_EN register is needed for the CPU port, any PHY as well as
any fixed-link.

This fixes a crash when trying to bring up a port which uses a fixed
link (in this case phydev in gswip_port_enable and
gswip_phylink_mac_link_up is NULL). Support for fixed-links will be
added in the future.

Fixes: 14fceff4771e51 ("net: dsa: Add Lantiq / Intel DSA driver for vrx200")
Cc: stable@vger.kernel.org
Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>