net: dsa: lantiq_gswip: Configure the MAC_CTRL0 register
For each switch port (excep the CPU port) there is a MAC_CTRL0 register
where various MAC related settings are configured. These include:
- xMII interface mode (GMII/RGMII or MII/RMII)
- Full Duplex
- Flow Control
- Transmit FCS

Explicitly elect "AUTO" mode for all of them when the link settings
are auto-negotiated by the PHY. This brings the registers into a defined
state when booting Linux, whereas before we relied on the bootloader to
configure them properly.

Fixes: 14fceff4771e51 ("net: dsa: Add Lantiq / Intel DSA driver for vrx200")
Cc: stable@vger.kernel.org
Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>