net: dsa: lantiq_gswip: Configure all remaining GSWIP_MII_CFG bits
There are a few more bits in the GSWIP_MII_CFG register for which we
did rely on the boot-loader (or the hardware defaults) to set them up
properly.

Previously in commit 3e6fdeb28f4c33 ("net: dsa: lantiq_gswip: Let GSWIP
automatically set the xMII clock") some improvements were made. However,
it turns out that some more changes are needed. Hauke reports that the
TRM for the xRX200 SoCs mentions: "MIIRATE: IN RMII mode, 50MHz must
always be selected" (so it seems that letting GSWIP choose the MII rate
does not work in case of RMII). Automatically let GSWIP set the MII rate
for all PHY modes except RMII - for RMII we manually configure 50MHz as
recommended by the TRM.

Also to fix the external AR8030 RMII PHYs on the Fritz 7360v2 we need to
select the GSWIP_MII_CFG_RMII_CLK bit.

The GSWIP IP also supports in-band auto-negotiation for RGMII PHYs. Set
or unset this bit depending on auto-negotiation mode.

While here, also add the GSWIP_MII_CFG_RESET bit. We don't need to
manage it because this bit is self-clearning when set. We still add it
here to get a better overview of the GSWIP_MII_CFG register.

Fixes: 14fceff4771e51 ("net: dsa: Add Lantiq / Intel DSA driver for vrx200")
Fixes: 3e6fdeb28f4c33 ("net: dsa: lantiq_gswip: Let GSWIP automatically set the xMII clock")
Cc: stable@vger.kernel.org
Suggested-by: Hauke Mehrtens <hauke@hauke-m.de>
Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>