net: dsa: lantiq_gswip: Add support for dumping the registers
Add support for .get_regs_len and .get_regs so it is easier to find out
about the state of GSWIP's auto polling mechanism and the decision the
hardware makes based on this data. For this we specifically add the
GSWIP_MAC_PSTATp(port) and GSWIP_MDIO_STATp(port) register #defines as
these contain the result of the auto polling mechanism. Other global
and per-port registers which are also considered useful are included as
well.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>