soc: amlogic: meson-mx-socinfo: Add support for the TrustZone firmware
When the TrustZone firmware is enabled the SoC is configured so the boot
ROM cannot be read from the (untrusted) Linux kernel. Instead a firmware
call needs to be used to get the SoC's "misc" version.
Add support for the firmware call to retrieve the SoC's misc version if
the TrustZone firmware is loaded.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>