dt-bindings: phy: meson8b-usb2: Add support for reading the ID signal
The first USB PHY on Amlogic Meson8/8b/8m2/GXBB SoCs is OTG capable.
This means that the USB "ID" signal is routed to the PHY. Add support
for the gpio-controller and #gpio-cells properties so the value of
the "ID" signal can be read as a GPIO (from the PHY) for example by
an "gpio-usb-b-connector".

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>