mtd: rawnand: mtk_ecc: Add support for the ECC on MT7621
ECC support on MT7621 is similar to existing SoCs. The main difference
is that this SoC does not support DMA based encoding/decoding and
requires fixing the bitflips (manually) in software instead. This also
results in some missing registers (which are DMA IRQ related). Apart
from that MT7621 only differs in having two fewer ECC strength settings
(compared to MT7622).

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>