phy: amlogic: Add a new driver for the CVBS DAC (CVBS PHY)
Amlogic Meson SoCs embed a CVBS DAC which converts the signal from the
VPU to analog. The IP has evolved over time with the SoC generations:
- Meson8/8b/8m2 has per-chip calibrated data for the CDAC_GSW register
- GXBB is overall similar to Meson8/8b/8m2 except that it doesn't have
  per-chip calibration data and uses 0x0 in CDAC_GSW always
- GXL/GXM are overall similar to GXBB but require a CDAC_VREF_ADJ value
  of 0xf (of which the actual meaning is unknown)
- G12A/G12B/SM1 use different register offsets and different values for
  CDAC_CTRL_RESV2, CDAC_VREF_ADJ and CDAC_RL_ADJ. Like other SoCs from
  GXBB onwards they don't need any per-chip data.

For backwards compatibility with old .dtbs the driver gets
platform_device_id's so the VPU driver can register the PHY as platform
device if not provided via .dtb.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>