clk: meson: mpll: consider sdm_en in .recalc_rate and .set_rate
The SDM value is the fractial part of an MPLL clock on Amlogic Meson
SoCs. This value is only used by the hardware when the sdm_en bit is set
to 1.

Make .recalc_rate read this bit and (depending on it's value) ignore or
honor the SDM value during rate calculation. Also update .set_rate to
toggle the bit depending on whether the SDM value is zero or non-zero.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>