mmc: meson-gx: Switch register access to an MMIO based regmap
Commit 8a38a4d51c5055 ("mmc: meson-gx: do not use memcpy_to/fromio for
dram-access-quirk") fixes an issue when the registers are read or
written with accessors that aren't 32-bit wide. According to a comment
from the vendor driver states that only 32-bit ABP (bus) access is
supported.

Switch the driver over to an MMIO based regmap to ensure that every
access is always 32-bit wide. This also simplifies the code in some
areas as regmap provides various helpers (e.g. regmap_update_bits
and regmap_bulk_{read,write}) that we can use.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>