net: dsa: lantiq_gswip: Fix FDB add/remove on the CPU port
There's no bridge available when adding or removing a static FDB entry
for (towards) the CPU port. This is intentional because the CPU port is
always considered standalone, even if technically for the GSWIP driver
it's part of every bridge.

While here return -EOPNOTSUPP in case we can't add/remove the FDB entry
for any currently unsupported scenario (for example when trying to add
or remove an FDB entry for a standalone port).

Handling the CPU port fixes the following message during boot:
  port 4 failed to add <LAN MAC address> vid 1 to fdb: -22
as well as the following message during system shutdown:
  port 4 failed to delete <LAN MAC address> vid 1 from fdb: -22

Fixes: 58c59ef9e930c4 ("net: dsa: lantiq: Add Forwarding Database access")
Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>