mtd: rawnand: intel: Prepare support for SoCs without HSNAND
Earlier Lantiq SoCs (whose NAND IP made it into the Intel LGM SoC) don't
have the HSNAND IP built-in. Boards with these SoCs only get support for
software or on-die (on NAND chip) ECC (as well as in theory "no ECC"
support, but no board is known to use it). Make support for the HSNAND IP
optional in the intel-nand-controller-driver in preparation for
supporting these older SoCs.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>