selftests: forwarding: lib: Move the ping destination IP argument last
musl's documentation states a difference in the getopt() behavior (which
is POSIX compatible) over the GNU (e.g. glibc) variant of getopt() [0]:
  GNU getopt permutes argv to pull options to the front, ahead of
  non-option arguments. musl and the POSIX standard getopt stop
  processing options at the first non-option argument with no
  permutation.

Move the destination IP address for the ping commands last. This makes
the ping and ping6 tests work on Linux distributions using musl libc
(such as OpenWrt) instead of glibc.

[0] https://wiki.musl-libc.org/functional-differences-from-glibc.html#Miscellaneous_functions_with_GNU_quirks

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>