selftests: forwarding: Use a wrapper function for mausezahn invocations
musl's documentation states a difference in the getopt() behavior (which
is POSIX compatible) over the GNU (e.g. glibc) variant of getopt() [0]:
  GNU getopt permutes argv to pull options to the front, ahead of
  non-option arguments. musl and the POSIX standard getopt stop
  processing options at the first non-option argument with no
  permutation.

Introduce a new wrapper function for mausezahn, similar to what we have
for ping and ping6. This wrapper function ensures that the interface
name (which is a non-option argument) is passed last. Also the "-q"
(quiet) parameter is moved to this wrapper function since all callers
are passing it in. Convert all direct $MZ callers to use this new
wrapper function to make tests using mausezahn work on Linux
distributions using musl libc (such as OpenWrt) instead of glibc.

[0] https://wiki.musl-libc.org/functional-differences-from-glibc.html#Miscellaneous_functions_with_GNU_quirks

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>