selftests: forwarding: Make mausezahn calls compatible with POSIX getopt
musl's documentation states a difference in the getopt() behavior (which
is POSIX compatible) over the GNU (e.g. glibc) variant of getopt() [0]:
  GNU getopt permutes argv to pull options to the front, ahead of
  non-option arguments. musl and the POSIX standard getopt stop
  processing options at the first non-option argument with no
  permutation.

Convert all remaining $MZ callers which are not using mz_do to pass the
interface name last (or second to last if an extra agument string is
given). This makes these tests using mausezahn work on Linux
distributions using musl libc (such as OpenWrt) instead of glibc.

[0] https://wiki.musl-libc.org/functional-differences-from-glibc.html#Miscellaneous_functions_with_GNU_quirks

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>