iio: adc: meson_saradc: Add support for the thermal sensor on more SoCs
GXBB, GXL, GXM and AXG SoCs also have a thermal sensor connected to the
SAR ADC (mux on internal channel 6 - just like Meson8/8b/8m2). There are
a few differences compared to the earlier SoCs:
- initialization of the thermal sensor ADC value (offset), TSC (thermal
  sensor coefficient) and calibration information now has to be read
  from an AO secure register (MESON_AO_SEC_SD_CFG12) instead reading it
  from the eFuse
- GXL, GXM and AXG SoCs additionally use a "thermal vref" (probably
  thermal voltage reference) value which also can be found in the same
  AO secure register, just like the other calibration information
- Dividers and multipliers vary with each SoC and GXL/GXM have a another
  difference where the divider depends on whether the thermal sensor's
  voltage reference is enabled or disabled

Implement SoC specific thermal sensor initialization functions for
managing the small differences between GXBB, GXL/GXM and AXG. The magic
numbers are based on the implementation from the vendor u-boot (in
common/cmd_cpu_temp.c).

Also implement logic to optionally set the thermal voltage reference
on GXL, GXM and AXG (according to the same vendor u-boot sources this
does not apply to GXBB or earlier SoCs).

The TSC (temperature sensor coefficient) register has moved compared to
Meson8/8b/8m2 from MESON_SAR_ADC_DELTA_10 (and the fifth bit of the TSC
in MESON_HHI_DPLL_TOP_0 on Meson8b and Meson8m2) to a new 5-bit wide
register in MESON_SAR_ADC_REG11.

To get meaningful values for the SoC temperature sensor the multipliers
and dividers are defined per SoC (again, also based on the vendor u-boot
implementation). The fixed offset of 27 stays the same for all SoCs and
so is the general formula. The formulas are (where "adc" is the raw ADC
value and "offset" is called temperature_sensor_adc_val in the
meson_saradc driver and based on the calibration values):
- Meson8: (18 * (adc - offset) * 10000) / 1024 / 10 / 85 + 27
- Meson8b/8m2: (10 * (adc - offset)) / 32 + 27
- GXBB: (10 * (adc - offset)) / 34 + 27
- GXL/GXM with thermal VREF enabled: (10 * (adc - offset)) / 153 + 27
- GXL/GXM with thermal VREF disabled: (10 * (adc - offset)) / 171 + 27
- AXG: (adc - offset) / 17 + 27

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>