iio: adc: meson_saradc: Rename tsc_regmap to hhi_regmap
The tsc_regmap field from struct meson_sar_adc_priv is only used on
Meson8b/8m2 SoCs where the fifth bit (BIT(4) when counting from zero) of
the TSC (temperature sensor coefficient) is part of the HHI registers.
On newer SoCs (starting with GXBB) the TSC is part of the SAR_ADC_REG11
register, which is part of the SAR ADC's own register space.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>