iio: adc: meson_saradc: Simplify enabling/disabling the bandgap
Enabling and disabling the bandgap depends on two factors:
- whether it's targeting the SoC temperature input channel (channel 6
  mux) or the voltage input channels
- the SoC generation

Enabling the bandgap when measuring the SoC temperature in the
MESON_SAR_ADC_DELTA_10 register is valid on all SoC generations where
the SAR ADC is used to measure the temperature (all SoCs up until AXG,
even if not yet implemented in the meson_saradc driver for GXBB, GXL,
GXM and AXG yet).

Enabling the bandgap when measuring a voltage input in the
MESON_SAR_ADC_REG11 register. This is only supported on GXBB and newer
because older SoCs don't have this register. Since we are using regmap
we can simply have a write to that register on older SoCs and the regmap
framework will convert this into a no-op, meaning we don't need any
extra conditions in the code.

Simplify enabling the bandgap by enabling the correct (for the channel
and channel type) bandgap setting. At the same time we can disable the
"other" bandgap setting. When disabling the hardware we disable all
bandgaps.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>