net: dsa: lantiq_gswip: Implement port_set_host_flood
The GSWIP hardware cannot enable flooding to the CPU port for each
individual user port. Instead flooding to the CPU port is managed
globally, meaning it's either enabled or disabled for all ports at the
same time.
Implement port_set_host_flood and keep the unicast/multicast flooding
active as long as any user port wants flooding to the host to be
enabled.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>