net: dsa: lantiq_gswip: Update per port flooding and learning defaults
Disable flooding and learning by default. These are sane defaults for
standalone (which the driver calls "single port bridge") ports. Apply
these settings during setup and while disabling the port. This ensures
that the settings are back to their defaults after removing a port from
a bridge and using the same port in standalone mode instead (OpenWrt's
netifd seems to leave flooding enabled when removing the port from a
bridge).

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>