net: dsa: lantiq_gswip: Manage the GSWIP_TABLE_VLAN_MAPPING valid bit
We manually have to set gswip_pce_table_entry.valid to true for the
hardware to actually use the entry. Not setting means the default value
is applied, which results in the entry being ignored.

Update gswip_vlan_add_{un,}aware() to mark the entry as valid as we want
it to be recognized by the hardware. Also update gswip_vlan_remove() to
mark the entry as invalid once all user ports have been removed (in this
case the GSWIP_TABLE_ACTIVE_VLAN entry is also removed later on, meaning
we want the hardware to forget about any of these entries), otherwise
mark the entry as valid.

Fixes: 8206e0ce96b3 ("net: dsa: lantiq: Add VLAN unaware bridge offloading")
Fixes: 9bbb1c053bdc ("net: dsa: lantiq: Add VLAN aware bridge offloading")
Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>