mtd: rawnand: meson: Remove explicit management of the "device" clock
The "device" clock is only used as parent of the NAND clock. This
however is specified by using nfc_divider_parent_data[0].fw_name =
"device", which means that the driver itself does not have to manage the
device clock. Common Clock Framework ensures that this parent clock is
automatically enabled and disabled based on it's consumer.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>